"1:20pm. 125/420. I am just skimming stuff at this point. I do not feel like reading all of this.
1:25pm. 138/420. This package management stuff is something I should revisit as I go along. I'll just take note of it for now.

1:35pm. 155/420. Let me just move on at the speed of light. I'll aim to finish the book in the next two hours and then I will reassess where I stand.

1:45pm. 181/420. Wow, there is a lot of stuff on packages here. Well, it is an important subject, but I'd like to just get on with it.

Let me stop for a bit so I could read Solo Leveling.

2pm. 199/420.

"Browsers are a crucial instrument in understanding complex systems or models. A browser is a tool to navigate and interact with a particular domain. Each problem domain is accompanied by an abundance of browsers that are created to help analyze and interpret the underlying elements. The issue with these browsers is that they are frequently (re)written from scratch, making them expensive to create and burdensome to maintain. While many frameworks exist to ease the development of user interfaces in general, they provide only limited support to simplifying the creation of browsers."

Oh, this chapter feels like it will be right up my alley. I'll read this one more carefully. I'll definitely be needing this for Spiral.

I'll also take a look at some of the videos by the guy who did RaptorJIT.

2:10pm. 217/420. Let me take a short break here.

I can't really go into depth on this right now.

...Actually let me continue. I want to just skim the entire book today. After that I will get to work on Spiral v0.1 with the intent on porting it to Pharo afterwards.

http://moosetechnology.org/

The videos on this might be worth watching.

2:30pm. http://stephane.ducasse.free.fr/FreeBooks/BlueBook/Bluebook.pdf
Smalltalk-80 - The Language And Its Implementation

...I spy a chapter on probability distributions. Now that is new for 36 years ago.

I won't read this now. It should be left for later.

https://stackoverflow.com/questions/37089768/does-pharo-provide-tail-call-optimisation

I got it from here. It seems Smalltalk does not do TCO.

https://discourse.julialang.org/t/trying-to-find-gui-options-for-julia-windows-linux/14575/14

Julia is Spiral's much hated rival so I am wondering where it stands on the whole GUI thing. Lightyears apart from Pharo it seems. Julia does have a fair bit of momentum behind it though so maybe it will get somewhere as a language. If my vision turns out to be right though, pretty much everything is going to relegated to a niche player apart from Pharo and Spiral.

3:40pm. Done with chores, let me resume. Let me commit this. After I am done with the book, I think I will take some time to compose the next PL monthly post. After that I will start real work. I'll redo Spiral for the ground up so as to wring out every last drop of inefficiency from it and after that I'll start taking Pharo and its interactivity seriously. It will be a long project, but I hope I will attain something by the end of it."