"4pm. 271/420. I think I am going to pay some attention to these chapters if possible. But to be honest, I really just want to get started on the monthly review so I can get it all out of my system.
I guess I'll do it tomorrow. I'll have plenty of time before it is time to leave for a bit.

4:15pm. 278/420. Here are the resumable exceptions at last.

5:15pm. Done with lunch. I want to get more into the book so I can spare myself of more of this tomorrow.

5:35pm. 297/420. "The mechanism is quite simple, making it worthwhile to know how it operates. On the contrary to most mainstream languages, exceptions are implemented in the language side without virtual machine support, using the reification of the runtime stack as a linked list of contexts (method or closure activation record)."

I want to read this, but I am pretty tired at this point.

Let me stop here for the day. The last quarter of the book seems like it will be worth studying.

...I guess I'll finish it tomorrow and then leave the review for after that."