Today I am taking a break so that the junk from yesterday's degustation can get cleaned from my system.
I finally got the GPU back into the rig - I had no idea that the GPU (and the mobo) had 4 different kinds of video ports. I thought the adapter was wrong at first. The old monitor's switched tricked me - I thought it was HDMI since that is what the mode I was using displayed, while it was DVI instead.

My plan for the rest of the day is to just watch anime, so I'll get back to it. Tomorrow I will resume the Deep into Pharo book.