"3:35pm. Finally done with the book.
https://www.youtube.com/watch?v=8pTEmbeENF4
Bret Victor The Future of Programming

I'll give this talk a chance, but besides that what I am going to do is take that monthly PL review and pare it down to a page or two of bullet points. Right now it is 4 pages long which is too much.

After that I begin programming. But I still feel like I haven't planned it it out enough. I will have regrets whichever way I go so I should just do what I want.

"
January was a deeply troubling month for me.

* I finally accepted the fact that I am only at 1/10th of my full productivity in [Spiral](https://github.com/mrakgr/The-Spiral-Language). It was different while it was still undergoing active development, but at this point I see this as absolutely unacceptable.

* I figured out a better internal representation for closures which would speed up the compiler significantly, but I do not want to deal with the drudgery of that anymore when I could be doing RL.

* I have a bunch of good ideas how to improve the language besides that - like Smalltalk's keyword arguments, but that would be even more work.

* "Ah, screw it. I might have worked on Spiral for 2 years just for ML, but let me just use Python + PyTorch. I'll get back to Spiral later when I need it."

* What I really want from Python is not so much PyTorch, but a deeply interactive experience which Spiral could not possibly offer me. I am deeply drained by the edit-compile-run cycle at this point.

* All the PyTorch examples have the plotting windows used for visualization frozen.

* The [Tkinter example](https://tkdocs.com/tutorial/firstexample.html) when run instead hangs the interpreter.

* The way to get around this requires messing with callbacks. This would push the effort needed to make graphical interfaces for my tools way up.

I just have no luck. I really want to do RL badly, but I want to also be complete as a programmer. In order to close the productivity gap Spiral needs much better tooling and performance, and not just that. Spiral has some deep inherent limitations owing to its static nature - despite its flexibility it is way out on the static side of the static/dynamic axis. It is the most static of static languages which is a poor fit for interactive development.

I've been thinking about what I could do with a real dynamic language on top of it, and given how simple Spiral is, it has great potential as a library language. It would serve well as a C++ killer for things like GPU and game programming.

At first I considered Python, but given that it would do very little to help me get away from the command line cycle I understand that for the sake of improving my future productivity I need to do something drastic. I've decided - I am going to get away from .NET, forget Python, forget RL and pick a top notch [interactive language](https://youtu.be/baxtyeFVn3w) as a target. I will make [Pharo](https://pharo.org/) my base.

By Looking into the future by extrapolating the present, some interesting conclusions can be drawn. Namely that if [these guys](https://youtu.be/cUTMhmVh1qs?t=1778) are going to make an attempt at causing the Singularity in the next decade, they are going to be doing it from the command line - using Python scripts.

I am not as smart as them, so I think it would be a good idea if I instead tried it by going up a different path. I'll pay the costs now so that I might have an easier time later. I think it would be a lot easier to make a ML library in Pharo than it would be to get Python's interactivity on Pharo's level. I quite like the idea of AI agents being a part of the VM image.

Live coding with objects will be that much more effective when the objects are literally alive.

So I guess I will be a regular in the PL monthly thread for the next year or two. I thought I was done with Spiral and that my last post would be my final one here, but I am pretty unlucky after all. Nobody else did the work that needs to be done nor seems likely to do it, so the responsibility for making that vision a reality falls onto me.
"

4:55pm. This is a much better review. It has the facts and it has the feeling. No need to burden the readers with every thought that comes to mind.

5pm. Let me take a look at the talk by Bret and then I'll call it a day. Tomorrow I will do some real programming."