Done with the break. Let me start.
This will be a real pain as I will have to redo not just the language, but also the libraries from the top. I am not sure whether all of this would really be worth it, but I need to do it in order to make the language faster at compile time. i need to break the dependency on matching on strings.

It needs to be done and I am the only one who cares enough about this to do it.

So let me do it.

Well, I should be done with it in about 3 weeks. Hopefully the Pharo stuff will be more fun.

Once I move to Pharo hopefully I will get some insight on what it really is by studying it in depth. This should give me insight on what dynamic languages really are.

I am really hoping that by the time all of this is really over, by the time the June update arrives I will have some real understanding of what a complete system really should be like.