"10:10am. Now I am just getting distracted. Yeah, I definitely did not think of this.
I sort of assumed that the situation on the module side would take care of itself but...

Consider that I am opening the core library before doing the other ones.

So something like

```
inl Core = ...
open Core
inl Option = ...
inl Tuple = ...
inl Liple = ...
```

That is how it was.

Now I have some issues.

The first of all would be that I cannot `open Core` until the prepass is done.

Another issue is that peval should be followed by the prepass for the next module in line. The reason for that is to allow `open` to work. For it to work it needs to partially evaluate the preceding module so it can calculate the stack size. But even though I could pass the environment, the one in the prepass does not have the name information in memory anymore.

Ack, what an oversight.

10:20am. Well there are two fixes that I can think of.

Add an extra module langauge at the top level.

```
inl Core = join_module ...
open Core
inl Option = join_module ...
inl Tuple = join_module ...
inl Liple = join_module ...
```

Add special join points for modules. That part will be necessary no matter what I decide.

10:25am. The other option is to not add a special language at the top level. Then I would do modules like so...

```
inl Core = join_module ...
join_module
    open Core
    inl Option = join_module ...
    inl Tuple = join_module ...
    inl Liple = join_module ...
```

In other words completely interleave all the passes.

10:30am. `join_module` would be special. It would not be evaluated during the prepass, but rather during partial evaluation itself. And what it would do is trigger the prepass and follow it up by partial evaluation.

The `RawExpr` version of `join_module` would be just a join point, but similarly to inlineables the `Expr` version would have `Map<int, string>` with which to reverse the renaming for the following pass. Actually it can just be `string []`.

10:40am. This is quite a quandary, but I am leaning towards the first idea.

The second one is more elaborate, but because of that what occurs to me is that...

```
inl Core = join_module ...
join_module
    open Core
    inl Option = join_module ...
    inl Tuple = join_module ...
    inl Liple = join_module ...
```

Should `Tuple` or some other inner module change, because how join points work that would trigger the recompilation of everything if I am operating under the second regime.

11am. Agh, I am going to go with the first idea. If I was really adventurous I'd go with the second, but I just know that this would cause me significant problems when it is time to port this to Pharo.

If I am expecting to recompile everything from scratch every time then the second idea is better. The first one is better if I actually expect to cache some things. And allowing caching is pretty much one of the main reason why I went for this thing.

11:05am. Right now, part of me is wondering whether I could have recursive modules. Supposing no `open` statements and no name collisions...

I am misunderstanding something important. The main reason why this sort of thing works for Pharo is because it is so simple. Simplicity has the great benefit of making other things simpler - in its case it is what allows mutual recursive class referencing.

Make Spiral an OO language in the vein of Smalltalk and I could get the same thing. But as a functional language, it has different sort of need doesn't it?

11:15am. Later when I am doing type debuggers I am going to need name information anyway.

11:20am. But no way am I going to carry the names around. That would defeat the entire point of doing all these optimizations. At most I should just add breakpoint nodes and infer the names in the environment from that.

...Let me have that brain here. My break is starting to overheat."