"12:15pm. Done with breakfast. I wasted it all on social sites.
Close that shit down. It sort of gets me in a mental state of anxiety where I am always looking for the next thing. It reminds me of the stock market.

I was not resting at all during the break.

The day after the part when I watched S4 and S5 RWBY without a care in the world was wonderful, but this sucks. Who cares that people at Davos are worried about the future. I should be working towards making them worried about me.

Let me resume for a while.

I think I might be bad at managing myself. I think I am good, but it is not like I have a reference for that.

Rather than being bad the programming state unlike games never truly fills me. It is a pity, but because of that I can never truly be the ideal programmer because of that.

But it does not matter. I am still pretty good, and now is the time to prove it. It feels me half the time, so I'll be satisfied with that.

12:25pm. I need to get my focus back.

12:30pm. I am super blocked up right now.

12:35pm. I am blocked up.

It is not as simple as having an outer language. Yeah, sure I can make a simple language that uses `Map<string,TypeData>` as the environment, but that would break the contextual link. Modules also need to be inlined.

12:40pm. I really want to just take a nap here, but let me put in a few thoughts.

Let me go back to the `join_module` idea. Who says that changing the inner module would necessitate a change in the outer ones?

Just why did that thought come to mind?

12:50pm. Let me take a nap. I need to get away from the screen for a bit.

This whole thing felt a lot more straightforward before I gave it a try.

If it was not for the modules I'd know how to do the whole language. Understand so many things, but just a few little pieces here are perplexing me."