"4:55pm. I could program, or I could surf /a/. Let me do it for another hour, programming I mean.
```
    and prepass (env: PrepassEnv) expr =
        let tag () = prepass_tag <- prepass_tag + 1; prepass_tag
        let prepass_memo_dict = Dictionary(HashIdentity.Reference)

        let rec loop env expr = memoize prepass_memo_dict (eval env) expr
        and eval env = function
            | RawV x -> V(tag(), string_to_var env x), 0

        loop env expr
```

Let me take care of the auxiliaries.

```
    let keyword_to_string _ = failwith ""
    let string_to_keyword _ = failwith ""
```

Let me take care of this. I will introduce the keyword dictionary.

5:20pm.

```
| If of Tag * cond: Expr * on_succ: Expr * on_fail: Expr
```

Let me also add the if as an explicit branch. I will need it in order to propagate stack size info through it properly.

```
    and prepass (env: PrepassEnv) expr =
        let tag () = tag_prepass <- tag_prepass + 1; tag_prepass
        let prepass_memo_dict = Dictionary(HashIdentity.Reference)

        let string_to_var env x =
            match Map.tryFind x env with
            | Some x -> x
            | None -> raise (PrepassUnboundVariable x)

        let rec loop (env: PrepassEnv) expr =
            memoize prepass_memo_dict (function
                | RawV x -> V(tag(), string_to_var env.prepass_map x), 0

                ) expr

        loop env expr
```

I really do not feel like continuing with this any longer.

I've been at it since 8am. Though the rate of progress has been disappointing, I did resolve how I want to do the modules themselves. Tomorrow I will aim to deal with the prepass and the pattern matching compiler.

5:30pm. Unlike the modules which I had not thought about in too much depth before today, the prepass is something that has been on my mind for the past month so I'll be able to make good progress once I catch wind of it. I'll blow through the code quickly once the flame gets lit.

Though I can't be sure, I think I might be able to move to testing in as little as a week and at most two weeks from now.

Why did I pick this path? In 2017 I kept pushing myself by saying that RL will be just around the corner, but now I can't really say that anymore. I'd really wished that my life were a simpler story where to gain success all I have to do is get better at programming. Spiral sort of failed on the open market, and the whole field of ML is not really helping me at the current juncture.

Am I going in the right direction or is there something that I lack?

If life were a simple story there would not be a need for emotions such as hate. Hate is not irrational and neither is spite.

For as long as I have desire, the work will get put in. I really do want to give Spiral proper internal representation at least before I move to Pharo which will be by the time February is done.

I think that once I move to Pharo I will take a break from doing this kind of work. I do not think I have the will to start Spiral for Pharo right away. Rather I think I will forget conventional RL for a bit and since Sandholm did deliver and made it so that counterfactual regret could be used to make a master level player on a desktop computer, I will play with that for a while.

Maybe to get a little taste of Pharo I will start things off by making a (graphical) poker game and a CFR-based agent for it.

5:45pm. It is absolutely horrible. From one perspective sure, I did make a language and did a lot of programming in it. But from another, having a great ML library seems even further away now than it did 9 months ago.

True greatness requires complete integration and a static language like Spiral cannot do it on its own.

Languages can Julia can sort of hack GPU programming even if not as well as Spiral, but literally nothing will come close to Pharo in interactivity.

If I could bring out the best of Spiral and the best of Pharo then...I will be able to win.

This will be so much work that I want to cry. From 2019 to 2020 I will be really doing hard, heavy training. Money is just a distant dream at this point. I am such a loser; maybe I am just incompetent at grasping that particular aspect of my goal?

5:55pm. Regardless, past is the past. I wanted a language in which a high quality ML library could be made with relative ease and succeeded at that task. Adding high interactivity as a goal will make things a lot more complicated, but reaching it push me to new heights as a programmer.

6:15pm. Done for the day. Let me have some fun."