"9:40am. Let me start. I had some ideas during the night.
```
| ObjectCreate of ObjectDict * FreeVars
```

I just realized something. I do not need the inner renaming step, I'd be better of using the whole outer environment as it is.

```
and ObjectDict = TaggedDictionary<KeywordArgTag,Expr * FreeVars * StackSize>
```

Let me get rid of `FreeVars` from here.

Oh, yes, I also forgot the case for creating keyword arguments...

Ah, no wait, actually I do not need that one. Strings can be converted to keywords during the...

But what will I use for ints...

Yeah, it would be just a slight optimization, but let me do it during the prepass after all.

```
| RawKeywordCreate of string * RawExpr []
...
| KeywordCreate of KeywordArgTag * Expr []
```

```
| IfStatic
```

Let me get rid of this from the language.

9:50am. Well, this should be enough to start off the day.

Actually, `ModuleWith` is also something I need. In the prepass. It is because of that `self` argument. I need to do the prepass analysis in order to take it into account."