I am up and brushing out the cobwebs. I really got up early today, too early in fact. Yawn.
Though I am just biding my time at the moment, it feels like I have some determination.

Reasoning, huh? I am really reflecting on various things. The stock market really has been annoying for the past 3 weeks. And Bayesian inference and deep learning have their own problems.

I really hate how difficult it is to reason anything out about deep learning. And yet, reasoning about reasoning is the whole challenge of the pre-Singularity era.

And Spiral despite me putting so much effort into it is still a neglected language. It is really a sad thing for it isn't it, that I am the only one that really believes in it.

I regret the 2007-2014 period, the 7 years that I spent trading, but I do not regret the last 4.

I regret that I let the stock market have such an easy bite out of me. 7 years is not a small amount of time for a human - a human cannot afford to waste it.

I now understand what true powerlessness is like. To just be pushed by the course of events because the intelect I possess is insufficient.

Everything I have and everything I posses will be scattered and wasted unless I succeed in causing the Singularity. If I had been born 100 years earlier, then I could accept that I have no choice but to live and die as a human, but the only thing I understand for certain is that in this era it is time to fight.

To live and die on my own terms.

Normals do not understand the games. Sometimes I get the feeling that they think that games themselves are normal, like another pair of shoes that they can buy in a shop somewhere.

That is not the end of this story. It can't be.

I won't let it be.

I know powerlessness, but I also know how to reason about reasoning. And that insight is telling me to keep going forward.

I will finish Spiral and do some more fun stuff in Pharo. Breaking the dependency on the command line has long been in the making and since things are like this, I might as well take the chance to do it now.

With RL, maybe I will try out the idea that I had, but to be honest, I want to give deep learning a pause. It is time that I gave CFR and maybe some Bayesian things a try.

I'll pick it up again once I start incorporating Spiral into Pharo.

I thought at some point I would stop working on languages, but I was wrong. Programming ability itself is about language creation.

No matter the age, once I decided to seriously pursue programming it became inevitable that I would make my own.

Even if by some miracle I get Spiral to the point where there is no more work to be done on it, by then I will have probably started work on some other language.

So that will be something I will have to accept and even enjoy.

I just know that moving away from all the type inference and command line stuff to a graphical environment will reinvigorate my programming.

I've been programming in the manner that I have for 4 years now and I had enough of it to last me a lifetime. Unless it is for money, I will never do this kind of thing again. I've been a fool - life is just too short to mess with scripts of any sort.

Scripts are not alive and will never be. They lack the persistence necessary for it.

...

It is time to start work on Spiral. The sooner I start, the sooner I finish the latest redesign.

Today I will finish the last few branches, the subrenaming and the pattern matching compiler.

Let me get on with it.