"9:30am. The first to come would be `RawOp`. The work I did yesterday is really beautiful. Hopefully I can in a similar effort today.
The sooner I finish Spiral the sooner I can begin anew. Once I am done with Spiral I will invent once more a new kind of programming. It is not just the lack of sleep talking - I am tired of programming like have. So much more should be possible. There are depths of integration I've not even touched upon.

Programming is not as weak as the thing I am doing now.

9:40am. Focus me, focus. You are just getting into the zone, do not start going to /a/ all of a sudden. You can do it.

...I am really getting a lot of mileage from `mapFold`. I did not use it nearly as much last time. This is just one of the ways in which I've grown.

```
                | RawOp(op,exprs) ->
                    let exprs, (free_vars, stack_size) =
                        Array.mapFold (fun (free_vars,stack_size) expr ->
                            let expr, free_vars', stack_size' = loop env expr
                            expr, (free_vars + free_vars', max stack_size stack_size')
                            ) (Set.empty, 0) exprs
                    Op(tag(),op,exprs),free_vars,stack_size
```

Here is `RawOp`. This one was easy. The next is `RawPos`. For it in particular there is no need to do anything special. I just need to trap exceptions and attach to them positional data.

9:50am.

```
                | RawExprPos(pos) ->
                    let pos, expr = pos.Pos, pos.Expression
                    try loop env expr
                    with
                    | :? PrepassError as er ->
                        let mes = er.Data0
                        raise (PrepassErrorWithPos(pos,mes))
```

I guess this will do for this particular branch.

9:55am. Now comes the subrenamer. The subrenamer is actually fairly simple unlike the main peval part, so it should not take more than 30m or so. I just have to iterate over everything and do the renaming. I also have to do memoization so as to not get piked by the pattern matcher.

...Before I start that let me check out something. Does Array.fold work on null arrays?"