Let me finally start for the day. I had better rest compared to yesterday, but making this new version of Spiral is stll a large strain on me.
In the next hour or so, let me finally finish the pattern matcher and then I'll see what comes after that. Yeah, I barely feel like doing the parser. I'd much rather get to the language itself.

Well, it will be done before I know it. Every stage has changes, but none of the changes are too huge so I'll be done with it eventually.

I want to finish this so I can finally have some fun with Pharo.

I think for the last thing I will do in Spiral before I leave it for a bit is test the exploration idea I outlined in the January PL post.