"1:50pm. Done with breakfast, chores and hygiene. It took me a while. I could have taken that bath yesterday after I was done with work, but I was too drained to bother with it.
Let me put in the `RawCase` and `Case` into the language and then I will move on to the parser.

1:55pm.

```
    | Let of Tag * bind: Expr * on_succ: Expr
    | Case of Tag * bind: Expr * on_succ: Expr
```

During the prepass at least, I think I will be able to reuse most of the code as for the let similarly to the tests.

Focus me. Let me start this."