"2pm.
```
| UnboxFoldLMap
```

Ah, there was this thing too. I found a use for this in the union module.

I'll find a way to make do without it this time. But boy I do not feel like doing all this programming. Redesigning the Union?

Agggghhhhh...

Let me just do the basic tests for the language and maybe I'll leave the rest for later. If I can get good results on the Parser test that will be enough for me.

2:10pm. The sheer prospect of doing all that heavy work froze me for a moment. Let me fix the `PatUnbox` branch.

2:15pm.

```
                    | PatUnbox pat ->
                        match pat with
                        | PatVar patvar -> RawCase(patvar,v arg,on_succ)
                        | _ -> let patvar = patvar() in RawCase(patvar,v arg,cp patvar pat on_succ on_fail)
```

Yeah, this is nice. This should work.

2:25pm. Focus me. The next comes the parser. Forget that unboxing fold. I will be able to replace it with a mapFold plus type equality checking for the one use case where it was needed."