"8:55am. I am up. Let me chill a little, maybe read Satanophany and then I'll start this.
9am. https://mangadex.org/title/32020/iron-ladies

He posted a chapter 4 hours ago. Does this guy ever sleep?

Iron Ladies is interesting. The author has zero sense of scale and the plot is about the MC fighting cartoon villians, but I've never seen a manga written in such a manner. And a MC like this would have never been written by a Japanese mangaka. The Chinamen seem to be much more about establishing their authority. I am impressed though that he has the chupzah to wank over China superiority when the angels are essentially using technology they do not understand that they took from other races.

Interesting parallel with the real world.

9:40am. Let me finally start. I've charged up my energy.

It is time to do the two keyword parsers and after that I'll be done with the most complex part of the compiler.

9:45am. Cracking credit assignment is the one overarching theme of the whole pre-Singularity era. Bayesian inference sort of skirts the whole issue of tackling it head on. But at some point the crack will form. Deep learning might not be able to go all the way, but it narrowed the possibilities quite a bit.

I've been too cocky with the stock market. In long term the stock market essentially does credit assignment. That means if I can crack CA then I can consider myself graduating from the school of speculation for good. I'll consider the shame of my failure to be wiped...though I'll still boil (((them))) alive anyway as an exercise.

9:50am. I am wondering in the absence of NNs whether regular compression methods like Huffman coding could be adapted to work with CFR.

I just give up on NNs. Gradient methods will not cut it for RL. SL is a beautiful story of improvement by making architectural innovations, but RL is just such a joke in terms of reliability. Deepmind is crushing Starcraft, but literally none of their research is any useful for increasing the reliability of training on the toy tasks.

I wish there was a Bayesian method that worked with minibatches.

The exploration method I invented over a month ago is interesting. Though it would not solve the reliability problems, I am actually impressed how easy it is to do nested inference using NNs.

There is a lot indication that differentiable programming research should be done in this direction. Reptile is just the start of it. I really wish I was the one that discovered Reptile - I actually thought of the idea of doing those kinds of updates long before the paper came out, but the issue is that I had absolutely no reason to think they would do anything, much less transfer learning.

In order to do this kind of programming and make any progress in this direction, I really need a better framework of how to reason about the whole optimization process. Else, the way I will discover things like Reptile would be by trying things out completely at random.

Furthermore though, the Bayesians have a point. Optimization should not be part of learning, at least not like this.

The story of gradient based optimization is far from over.

10:05am. Let me finally start for real.

Doing these two parsers won't be hard."