If I am going to fail, at the very least I should make sure that I accumulate useful skills along the way. I could not accept it, but as soon as I set out on this path, I was fated to lose. It is unfair that I have had to lose in such a manner over the past decade and a half, but the killer-machines I will make for the sake of retribution won't be fair either.
Programming is wonderful. I completely failed to meet my goals in 2018, but 4 more years of failures such as this and I will be the best in the world at programming.

High school and then trading in contrast just wasted a decade of my life for no reason. This sort of failure in contrast is acceptable.

Let me chill a little and then I will start.