I can only be ashamed that I work so hard all the time for so little immediate gain. It is time for another day of that. I'll get into it once I start.
Time can lie.

Consider how a casino might win on most days despite having a trivial 0.5% edge. That tiny difference is the difference between life and death.

To humans who do not have the ability to change talents, those kinds differences are something absolute. Those cannot overcome nature will inevitably be overcome by it.

Sure, I want my billions and I want my harem and godlike power. I want it all.

Ironically even though I do all of this, what I am trying to attain by programming is not something huge.

What I want is just a little bit of reliability, that tiny 1% that has been eluding me.

If I can grasp the next step in ML and gain a little bit of that blessing, I can ride the wave all the way to its completion.

It has been quite hard. Despite putting in such a huge effort over the past year I haven't been able to do it. The others cannot do it either.

Deepmind cannot do it right now. OpenAI cannot do it.

It is fine to go at it like this - to gradually build on my work every day and improve slowly and steadily in such a manner.

Money, friends, family - had I power I would not need to make difficult choices, but since I do not I need to go forward bearing the risks that one day the journey might be cut off before it reaches its climax.

I suppose I can only be envious of those currently making it big using supervised learning.

But the only thing I really regret is that I had not started this in 2004 instead of 2014. Had I done that, my position right now would have been so much better.

I was unfortunate that I had no reason to start in 2004. In a way Simulacrum was a gift to the me of the past. It is similar to how the current Spiral documentation is written for the me of 2016. I am sure that if I had seen a true story of self improvement such as Simulacrum, I would have understood the message behind it in 2004.

I would have understood that some things are simply impossible to write sanely and things that cannot be sane will necessarily have to be insane.

Right now it is year 2019. The time I will have spent programming today will be a gift to the me of 2029.

So that in the head of the battle when he falters and inevitably looks back, he will see my figure pushing him forward. It will give him the strength to keep fighting. He cannot do it alone.

Let me start. I can still become so much stronger.

If I can master dynamic languages beyond the surface level, it will be an incredibly strong power. Before that I first must finish Spiral. It is out of the question to leave it in a half baked state.