"4:50pm. Judging by the number of lines, there are around 100 cases that I have to implement. Yeah, no way I am going through that in 2 days. More like 7.
```
| Apply,[a;b] -> apply_tev d a b
| IfStatic,[cond;tr;fl] -> if_static d cond tr fl
| While,[cond;body] -> while_ d cond body
| JoinPointEntryMethod,[a] -> join_point_method d a
| JoinPointEntryType,[a] -> join_point_type d a
| JoinPointEntryCuda,[a] -> join_point_cuda d a
| TermCast,[a;b] -> term_cast d a b
| LayoutToNone,[a] -> layout_to_none d a
| LayoutToStack,[a] -> layout_to LayoutStack d a
| LayoutToPackedStack,[a] -> layout_to LayoutPackedStack d a
| LayoutToHeap,[a] -> layout_to LayoutHeap d a
| LayoutToHeapMutable,[a] -> layout_to LayoutHeapMutable d a

| Dynamize,[a] -> dynamize d a
| (MacroFs | MacroCuda),[a;b] -> macro op d a b
| (CudaTypeCreate | DotNetTypeCreate), [a] -> extern_type_create op d a
| TypeAnnot,[a;b] -> type_annot d a b
| TypeUnion,l -> type_union d l
| TypeBox,[a;b] -> type_box d a b
| TypeGet,[a] -> type_get d a
| TypeCatch,[a] -> type_catch d a
| TypeRaise,[a] -> type_raise d a
| TypeSplit,[a] -> type_split d a
| TermFunctionTypeCreate,[a;b] -> term_fun_type_create d a b
| TermFunctionDomainRangeCPS,[a;b;c] -> term_fun_dom_range_cps d a b c
```

5pm. I've made my plan for tomorrow. I will take care of the branching, the function application, join points, term casting and layout types. Those are Spiral's core features. Then I will deal with all the type stuff.

The rest after that are primitives which I will be to easily reuse from the previous partial evaluator. I will be able to go through those quickly and I do not have much in the way of changes in mind for that.

5:05pm. 3 days seems to be a reasonable estimate for when I will be all done with the pevaller. I should be mostly done after 2.

5:10pm. Nnnnhhhh, I had enough for today.

Let me just say one thing, I am beyond happy that once I am done implementing all of this that I won't have to deal with type errors like in Spiral. If I've been doing this in Spiral, just the type errors would case it take a week or two longer for the entire project to finish. Maybe even more. It would be hell.

Imagine spending like a month fixing type errors. Brrrrrr.

This is something I need to get away from.

5:15pm. Since I am at it, it might be interesting to take Pharo and add a ML-like DSL if I am going to be doing a lot of language creation in it. Compared to two years ago, I am not so against classical type inference schemes.

5:20pm. Back then I was not sure I could even do it, but after the two year Spiral gauntlet, I can't imagine there being a programming task I cannot tackle. Solving equality constraints should not be that bad compared to making a partial evaluator.

Those tasks are all something I will have to tackle eventually.

After I am done with the pevaller, I will take a break for a bit to go through the CFR Minization tutorial. I want to figure out whether those methods are something I could potentially do on the GPU. There is a bunch of Java code in the paper.

Well, it is just a thought. Maybe I could implement the CFR examples in Spiral as a way of testing the new implementation?

5:25pm. I have a ton of work over the next 4 weeks waiting for me. It is going to be a real marathon and I am already exhausted. But that is how it always was. There was not a single day in the last 4 years since I set out on this path that was easy.

I'll get some rest and get ready to tame the chaos of reality once again after that.

I won't be held back forever by this obstacle. The work I did in the past 9 months in ML will come to life once I figure out how to do proper hierarchies rather than just tacking layers on top of each other. Me sense is that there is a whole dimension that is currently eluding my and community's understanding of machine learning. My sense is that nested inference is the key to unlocking the true power of machines, more specifically, I think it will turn out less intractable than it seems at first glance.

The ML library itself that I made for the sake of going in this direction is not enough. It is fine for simple nets, but for doing proper research I need something that will allow me to move much faster.

I need to make a much better library, in a language more suited for it.

Not something that is just fast, but something that is complete in every way.

Abstraction is the only true friend I have on this journey. Abstractions build on each other and accumulate in power without end. I will believe in them. They are what will allow some of my meager faculties to compete and eventually surpass the environment that created me."